requirements.txt
datasets==2.18.0
pyarrow==15.0.0
kagglehub
torch
torchvision
pillow
wandb